Added postgres database for production. Replaced JdbcTemplate custom repository by jpa repository.

+++ src/main/java/pt/fvaz/koerber/challenge/repository/TopZone.java
@@ -3,5 +3,5 @@ package pt.fvaz.koerber.challenge.repository;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record TopZone(@JsonIgnore Integer zoneId, String zone, @JsonProperty("pu_total") Integer puTotal, @JsonProperty("do_total") Integer doTotal) {
+public record TopZone(@JsonIgnore Long zoneId, String zone, @JsonProperty("pu_total") Long puTotal, @JsonProperty("do_total") Long doTotal) {
 }
